+++ src/app/(disputes)/create/page.tsx
@@ -2,7 +2,6 @@
 
 import React, { useState } from "react";
 import { useRouter } from "next/navigation";
-import { useQueryClient } from "@tanstack/react-query";
 import { toast } from "sonner";
 import {
   ArrowLeft,
@@ -15,8 +14,7 @@ import {
   ShieldAlert,
 } from "lucide-react";
 
-import { useCreateDispute } from "@/hooks/useCreateDispute";
-import { uploadFileToIPFS } from "@/util/ipfs";
+import { useCreateDisputeForm } from "@/hooks/useCreateDisputeForm";
 import { Button } from "@/components/ui/button";
 
 // Import Modular Components
@@ -28,8 +26,6 @@ import {
   StepReview,
 } from "@/components/create";
 import type {
-  CreateDisputeForm,
-  FileState,
   StepDefinition,
 } from "@/components/create";
 
@@ -43,45 +39,22 @@ const STEPS: StepDefinition[] = [
 
 export default function CreateDisputePage() {
   const router = useRouter();
-  const queryClient = useQueryClient();
-  const { createDispute, isCreating } = useCreateDispute();
+
+  // --- CUSTOM HOOK ---
+  const {
+    formData,
+    updateField,
+    files,
+    setFiles,
+    submit,
+    isProcessing
+  } = useCreateDisputeForm();
 
   // --- WIZARD STATE ---
   const [currentStep, setCurrentStep] = useState(1);
-  const [isUploading, setIsUploading] = useState(false);
   const [showDefenderOptions, setShowDefenderOptions] = useState(false);
 
-  // --- FORM STATE ---
-  const [formData, setFormData] = useState<CreateDisputeForm>({
-    title: "",
-    category: "General",
-    jurorsRequired: 3,
-    deadlineHours: 96,
-    claimerName: "",
-    claimerAddress: "",
-    defenderName: "",
-    defenderAddress: "",
-    description: "",
-    evidenceLink: "",
-    defDescription: "",
-  });
-
-  // --- FILE STATE ---
-  const [files, setFiles] = useState<FileState>({
-    audio: null,
-    carousel: [],
-    defAudio: null,
-    defCarousel: [],
-  });
-
   // --- HANDLERS ---
-  const updateField = (
-    field: keyof CreateDisputeForm,
-    value: string | number,
-  ) => {
-    setFormData((prev) => ({ ...prev, [field]: value }));
-  };
-
   const handleNext = () => {
     // Basic Validation per step
     if (currentStep === 1 && !formData.title)
@@ -98,92 +71,6 @@ export default function CreateDisputePage() {
     if (currentStep > 1) setCurrentStep((c) => c - 1);
     else router.back();
   };
-
-  // --- SUBMISSION LOGIC ---
-  const handleSubmit = async () => {
-    if (formData.jurorsRequired % 2 === 0) {
-      toast.error("Please select an odd number of jurors.");
-      return;
-    }
-
-    try {
-      setIsUploading(true);
-
-      // 1. Upload Claimant Assets
-      let audioUrl = "";
-      if (files.audio) {
-        toast.info("Uploading claimant audio...");
-        const hash = await uploadFileToIPFS(files.audio);
-        if (hash) audioUrl = `https://gateway.pinata.cloud/ipfs/${hash}`;
-      }
-
-      let carouselUrls: string[] = [];
-      if (files.carousel.length > 0) {
-        toast.info("Uploading claimant photos...");
-        const uploadPromises = files.carousel.map((f) => uploadFileToIPFS(f));
-        const hashes = await Promise.all(uploadPromises);
-        carouselUrls = hashes
-          .filter((h) => h)
-          .map((h) => `https://gateway.pinata.cloud/ipfs/${h}`);
-      }
-
-      // 2. Upload Defender Assets
-      let defAudioUrl: string | null = null;
-      if (files.defAudio) {
-        const hash = await uploadFileToIPFS(files.defAudio);
-        if (hash) defAudioUrl = `https://gateway.pinata.cloud/ipfs/${hash}`;
-      }
-
-      let defCarouselUrls: string[] = [];
-      if (files.defCarousel.length > 0) {
-        const hashes = await Promise.all(
-          files.defCarousel.map((f) => uploadFileToIPFS(f)),
-        );
-        defCarouselUrls = hashes
-          .filter((h) => h)
-          .map((h) => `https://gateway.pinata.cloud/ipfs/${h}`);
-      }
-
-      // 3. Construct Payload
-      const disputePayload = {
-        title: formData.title,
-        description: formData.description,
-        category: formData.category,
-        evidence: formData.evidenceLink ? [formData.evidenceLink] : [],
-        aliases: {
-          claimer: formData.claimerName || "Anonymous Claimant",
-          defender: formData.defenderName || "Anonymous Defendant",
-        },
-        audioEvidence: audioUrl || null,
-        carouselEvidence: carouselUrls,
-        defenderDescription: formData.defDescription || null,
-        defenderAudioEvidence: defAudioUrl,
-        defenderCarouselEvidence: defCarouselUrls,
-        created_at: new Date().toISOString(),
-      };
-
-      const success = await createDispute(
-        formData.defenderAddress,
-        formData.claimerAddress || undefined,
-        formData.category,
-        disputePayload,
-        formData.jurorsRequired,
-        formData.deadlineHours,
-      );
-
-      if (success) {
-        await queryClient.invalidateQueries({ queryKey: ["disputeCount"] });
-        router.push("/profile");
-      }
-    } catch (err) {
-      console.error(err);
-      toast.error("Failed to upload evidence.");
-    } finally {
-      setIsUploading(false);
-    }
-  };
-
-  const isProcessing = isCreating || isUploading;
 
   // --- RENDER CURRENT STEP ---
   const renderStep = () => {
@@ -250,7 +137,7 @@ export default function CreateDisputePage() {
       {/* --- FLOATING FOOTER --- */}
       <div className="absolute bottom-0 left-0 right-0 p-6 bg-gradient-to-t from-white via-white/95 to-transparent z-30">
         <Button
-          onClick={currentStep === 4 ? handleSubmit : handleNext}
+          onClick={currentStep === 4 ? submit : handleNext}
           disabled={isProcessing}
           className={`
             w-full py-6 rounded-2xl font-manrope font-bold text-base shadow-xl
